+++ CHANGELOG.md
@@ -32,6 +32,12 @@ This document details all notable changes to the WSU HRS Child Theme. Uses [Sema
 - Get Babel processing working, open #35.
 - Determine if we really need the `center-middle` utility class. If yes, look into using a flex or grid with `Xvh` and `Xvh` column and row size to center instead.
 
+## 0.13.1 (2018-06-08)
+
+### Fixed
+
+- HRS child theme JS no longer uses jQuery, so remove that as a dependency.
+
 ## 0.13.0 (2018-06-08)
 
 ### Added